+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 - Split ``unbox`` subcommand into ``join`` and ``boilerplate``.
 - Less verbose output
 - Simplified uwsgi config (vassal_defaults)
+- Fixed ``config`` command
 
 0.1 (25.06.2013)
 ----------------